requirements.txt
fastapi
uvicorn
pytest
httpx
python-dotenv
SQLAlchemy
psycopg2-binary
